Refactored server files
